LrcFile: cache {cpu,flow}_evaluation_data if delete_loaded_data=True
When user wants to use {cpu,flow}_evaluation_data properties, the delete_loaded_data
must be set to True. However, the content of these properties are small
dictionaries with data parsed out of the `_evaluation_data`.

To be able to use these properties also with delete_loaded_data=True, we
need to cache them while initializing the LrcFile instance.

Signed-off-by: Jan Tluka <jtluka@redhat.com>